6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,11 +6,10 @@
 int main(void)
 {
 
-printf("size of char : %d bytes\n", (unsigned long)sizeof(char));
-printf("size of int : %d bytes\n", (unsigned long)sizeof(int));
-printf("size of long int : %d bytes\n", (unsigned long)sizeof(long int));
-printf("size of long long int : %d bytes\n", (unsigned long)
-		sizeof(long long int));
-printf("size of float : %d bytes\n", (unsigned long)sizeof(float));
+printf("size of char : %d bytes\n", sizeof(char));
+printf("size of int : %d bytes\n", sizeof(int));
+printf("size of long int : %d bytes\n", sizeof(long int));
+printf("size of long long int : %d bytes\n", sizeof(long long int));
+printf("size of float : %d bytes\n", sizeof(float));
 return (0);
 }
